+++ pandora_console/operation/agentes/estado_agente.php
@@ -114,7 +114,7 @@ ob_end_clean();
 // Take some parameters (GET)
 $group_id = (int) get_parameter ("group_id", 0);
 $search = trim(get_parameter ("search", ""));
-$offset = get_parameter('offset', 0);
+$offset = (int)get_parameter('offset', 0);
 $refr = get_parameter('refr', 0);
 $recursion = get_parameter('recursion', 0);
 $status = (int) get_parameter ('status', -1);
